+++ packages/learn-card-core/src/types/LearnCard.ts
@@ -2,9 +2,11 @@ import { ModelAliases } from '@glazed/types';
 import { VerificationItem, UnsignedVC, VC, VP, VerificationCheck } from '@learncard/types';
 
 import { DidMethod } from '@wallet/plugins/didkit/types';
-import { DidKeyPluginMethods } from '@wallet/plugins/didkey/types';
+import { Algorithm, DidKeyPluginMethods } from '@wallet/plugins/didkey/types';
+import { EthereumPluginMethods } from '@wallet/plugins/EthereumPlugin/types';
 import { IDXCredential, IDXPluginMethods } from '@wallet/plugins/idx/types';
 import { VCPluginMethods } from '@wallet/plugins/vc/types';
+import { EthereumConfig } from '@wallet/plugins/EthereumPlugin/types';
 import { InitInput } from '@didkit/index';
 
 import { UnlockedWallet } from 'types/wallet';
@@ -12,8 +14,8 @@ import { UnlockedWallet } from 'types/wallet';
 export * from '@learncard/types';
 
 export type LearnCardRawWallet = UnlockedWallet<
-    'DIDKit' | 'DID Key' | 'VC' | 'IDX' | 'Expiration',
-    DidKeyPluginMethods<DidMethod> & VCPluginMethods & IDXPluginMethods
+    'DIDKit' | 'DID Key' | 'VC' | 'IDX' | 'Expiration' | 'Ethereum',
+    DidKeyPluginMethods<DidMethod> & VCPluginMethods & IDXPluginMethods & EthereumPluginMethods
 >;
 
 export type LearnCardWallet = {
@@ -25,7 +27,7 @@ export type LearnCardWallet = {
     /** Wallet holder's did */
     did: (type?: DidMethod) => string;
     /** Wallet holder's ed25519 key pair */
-    keypair: { kty: string; crv: string; x: string; d: string };
+    keypair: (type?: Algorithm) => { kty: string; crv: string; x: string; y?: string; d: string };
 
     /** Signs an unsigned Verifiable Credential, returning the signed VC */
     issueCredential: (credential: UnsignedVC) => Promise<VC>;
@@ -79,6 +81,21 @@ export type LearnCardWallet = {
      * You can use this to test out implementations that use this library!
      */
     getTestVc: (subject?: string) => UnsignedVC;
+
+    /**
+     * Returns your ETH balance
+     */
+    checkMyEth: () => Promise<string>;
+
+    /**
+     * Returns your DAI balance
+     */
+    checkMyDai: () => Promise<string>;
+
+    /**
+     * Returns your USDC balance
+     */
+    checkMyUsdc: () => Promise<string>;
 };
 
 export type CeramicIDXArgs = {
@@ -92,4 +109,5 @@ export type LearnCardConfig = {
     ceramicIdx: CeramicIDXArgs;
     didkit: InitInput | Promise<InitInput>;
     defaultContents: any[];
+    ethereumConfig: EthereumConfig;
 };