+++ packages/learn-card-core/src/wallet/defaults.ts
@@ -1,4 +1,5 @@
 import { CeramicIDXArgs } from 'types/LearnCard';
+import { EthereumConfig } from './plugins/EthereumPlugin/types';
 
 export const defaultCeramicIDXArgs: CeramicIDXArgs = {
     modelData: {
@@ -17,4 +18,10 @@ export const defaultCeramicIDXArgs: CeramicIDXArgs = {
     credentialAlias: 'MyVerifiableCredentials',
     ceramicEndpoint: 'https://ceramic-node.welibrary.io:7007',
     defaultContentFamily: 'SuperSkills',
+};
+
+export const defaultEthereumArgs: EthereumConfig = {
+    address: '',
+    infuraProjectId: '',
+    network: 'mainnet',
 };