+++ packages/learn-card-bridge-http/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @learncard/create-http-bridge
 
+## 1.0.6
+
+### Patch Changes
+
+-   Updated dependencies [[`4028716`](https://github.com/WeLibraryOS/LearnCard/commit/40287160de54d06f7baff000dee6f59f08f8623a)]:
+    -   @learncard/core@1.5.1
+
+## 1.0.5
+
+### Patch Changes
+
+-   Updated dependencies [[`de4e724`](https://github.com/WeLibraryOS/LearnCard/commit/de4e7244961f0ef91b91e6cbf32a43f29ff58b96)]:
+    -   @learncard/core@1.5.0
+
+## 1.0.4
+
+### Patch Changes
+
+-   Updated dependencies [[`e72b559`](https://github.com/WeLibraryOS/LearnCard/commit/e72b55994495e4bc6156b08abdd166c77fae67b7)]:
+    -   @learncard/core@1.4.0
+
 ## 1.0.3
 
 ### Patch Changes